% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hypothesis_tests.R
\name{retro_design_closed_form.list}
\alias{retro_design_closed_form.list}
\title{List retro_design_closed_form}
\usage{
\method{retro_design_closed_form}{list}(A, s, alpha = 0.05)
}
\arguments{
\item{A}{a list, estimates of the true effect size}

\item{s}{a numeric, standard error of the estimate}

\item{alpha}{a numeric, the statistical significance threshold}
}
\value{
A df that is 4 by length(A), with an effect size
and it's corresponding power, type s, and type m errors in each row.
}
\description{
retro_design_closed_form.list is the S3 method of the generic retro_design_closed_form() function,
used when a list is passed for A.
}
\examples{
retro_design(list(.2,2,20),8.1)
}
