% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hypothesis_tests.R
\name{retro_design_closed_form.numeric}
\alias{retro_design_closed_form.numeric}
\title{Numeric retro_design_closed_form}
\usage{
\method{retro_design_closed_form}{numeric}(A, s, alpha = 0.05)
}
\arguments{
\item{A}{a numeric, an estimate of the true effect size}

\item{s}{a numeric, standard error of the estimate}

\item{alpha}{a numeric, the statistical significance threshold}
}
\value{
A list of length 3 containing the power, type s, and type M
error.
}
\description{
retro_design_closed_form.numeric is the S3 method of the generic retro_design_closed_form() function,
used when a single numeric is passed for A.
}
\examples{
retrodesign(1,3.28)
retrodesign(2,8.1)
}
