% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hypothesis_tests.R
\name{retrodesign}
\alias{retrodesign}
\title{Calculates Power, Type S, and Type M error and returns them in a list.
Function originally provided in Gelman and Carlin (2014), reused with
permission under the MIT license.}
\usage{
retrodesign(A, s, alpha = 0.05, df = Inf, n.sims = 10000)
}
\arguments{
\item{A}{a number, an estimate of the true effect size}

\item{s}{a number, standard error of the estimate}

\item{alpha}{a number, the statistical significance threshold}

\item{df}{a number, the degrees of freedom}

\item{n.sims}{a number, how many times to simulate when calculating Type M
error}
}
\description{
Calculates Power, Type S, and Type M error and returns them in a list.
Function originally provided in Gelman and Carlin (2014), reused with
permission under the MIT license.
}
\examples{
retrodesign(1,3.28)
retrodesign(2,8.1)
}
