% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hypothesis_tests.R
\name{type_s.list}
\alias{type_s.list}
\title{List type_s}
\usage{
\method{type_s}{list}(A, s, alpha = 0.05)
}
\arguments{
\item{A}{a list, estimates of the true effect size}

\item{s}{a numeric, standard error of the estimate}

\item{alpha}{a numeric, the statistical significance threshold}
}
\value{
A df that is 2 by length(A), with an effect size
and it's corresponding type s errors in each row.
}
\description{
type_s.list is the S3 method of the generic type_s() function,
used when a list is passed for A.
}
\examples{
type_s(list(.2,2,20),8.1)
}
