% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hypothesis_tests.R
\name{type_s.numeric}
\alias{type_s.numeric}
\title{Numeric type_s}
\usage{
\method{type_s}{numeric}(A, s, alpha = 0.05)
}
\arguments{
\item{A}{a numeric, estimate of the true effect size}

\item{s}{a numeric, standard error of the estimate}

\item{alpha}{a numeric, the statistical significance threshold}
}
\value{
either the type S, a numeric if a single A is provided, or a df
of length 2 by A, with the effect size and corresponding type S error in
each row.
}
\description{
this is the S3 method of the generic type_s() function,
used when a numeric is passed for A.
}
\examples{
type_s(1,3.28)
}
