% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hypothesis_tests.R
\name{type_s}
\alias{type_s}
\title{type_s}
\usage{
type_s(A, s, alpha = 0.05)
}
\arguments{
\item{A}{a numeric or list, estimate(s) of the true effect size}

\item{s}{a numeric, standard error of the estimate}

\item{alpha}{a numeric, the statistical significance threshold}
}
\value{
either the type S, a numeric if a single A is provided, or a df
of length 2 by A, with the effect size and corresponding type S error in
each row.
}
\description{
Calculates type s error.
}
\examples{
type_s(1,3.28)
type_s(list(.2,2,20),8.1)
}
